requirements.txt
click==3.3
ghalton==0.6
-e git+https://github.com/benanne/Lasagne.git@9f591a5f3a192028df9947ba1e4903b3b46e8fe0#egg=Lasagne-master
matplotlib==1.4.3
-e git+https://github.com/dnouri/nolearn.git@0a225bc5ad60c76cdc6cccbe866f9b2e39502d10#egg=nolearn-master
numpy==1.10.4
pandas==0.16.0
Pillow==2.7.0
PyYAML==3.11
scikit-image==0.11.3
scikit-learn==0.17
scipy==0.16.1
SharedArray==0.2
Theano==0.8
joblib==0.9.3
tabulate==0.7.5
